Refactoring
Moved service to another location under /generated/.
